refactor(renderer): extract ToolCallSummary and reuse in streaming view
